backend_api/correlation_engine/requirements.txt
fastapi
uvicorn
kafka-python
httpx
psycopg2-binary